Find associated JSDoc when preceding a named export
The JSDoc comment defining a namespace etc. assigned to a class is found
when preceding a default export statement which wraps the class
definition. But it is so far not found when the export statement is a
*named* export.

Why is this required? For custom control development, UI5 is about to
provide a tool that generates the method declarations for the methods
generated at runtime. The methods are generated as an interface in a
separate file. Due to equal names, TypeScript merges the interface with
the class. But the (control) class must also be a named export in
addition to the regular default export, in order to make this mechanism
work. Hence, controls shall be written like this:

/**
 * @namespace x.y
 */
export class MyControl extends Control {
 ...
};
export default MyControl;

This change allows the preceding JSDoc to be found and considered.