Introduce JSDoc annotation denoting named exports as design-time only
Using a JSDoc comment with the newly introduced custom tag
@ui5designtimeonlyexport the subsequent named export can be marked as
"design-time only". This means it is only needed during development time
for certain TypeScript mechanisms, but should not be present at runtime.

This implements a proposal discussed in
https://github.com/r-murphy/babel-plugin-transform-modules-ui5/pull/66
to make the overall solution discussed in
https://github.com/SAP/ui5-typescript/issues/294
prettier:
For module merging needed in UI5 control development, the control class
name needs to be a named export (in addition to being the default
export). This is only needed at development time to make the TypeScript
compiler recognize the additionally generated interface as enriching the
control class.