Add "overrides" to static props which go into the extend(...) object
This is needed for ControllerExtensions: the "overrides" block needs to
go into the ControllerExtension.extend() call, not appended to the class
later as a static property.

Note that the name for this definition block is still "override" in the
documentation at this time. But this is about to change, because a
static property "override" clashes with the method
ControllerExtension.override() in TypeScript.

Contributes to fixing https://github.com/SAP/ui5-typescript/issues/332