fix: support index modules for dependencies
CommonJS or ES modules require index modules when a folder is
required, e.g.:

```js
import * from "./schema"
```

will in case of `./schema` is a directory lookup the index module
and require the following:

```js
import * from "./schema/index"
```

under the hood. As the UI5 module loader cannot support this feature
this Babel plugin checks for the existence of the index module and
rewrites the import to the index module so that the UI5 loader also
loads the proper module.