feat: support ControllerExtensions extended with .override(...)
Recognize (and remove) a dummy function ControllerExtension.use(...)
that only exists on TypeScript/ES-class side and takes care of
converting the class resulting from SomeExtension.override(...) to the
instance type that needs to be presented to TypeScript for the member
property.

Intentionally does not remove the previous @transformControllerExtension
annotation to not disrupt potential users immediately. But it will be
removed.

Contains some unrelated whitespace fixes for prettier.