fix: move qunit config autostart out of require/define
In case of using QUnit.config.autostart to configure the globally
available QUnit, it must be moved out from the sap.ui.require or
sap.ui.define block so that it is applied synchronously when loading
the module.

If QUnit is required locally, then QUnit.config.autostart will not
be moved.

The move can also be supressed by using the configuration option
noWrapQUnitConfigAutostart.